Add support for Gradle Init Scripts written in Kotlin
When versioned settings is enabled the DSL compiler will try to compile any
file with a kts extension found in the project directory. This includes files
stored in the plugin data directory. To avoid the DSL compiler failing and
to support Kotlin init scripts requires encoding and decoding the
kts extension.